# server/public/openApi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/system/role": {
      "post": {
        "operationId": "RoleController_create",
        "summary": "角色管理-创建",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "角色管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      },
      "put": {
        "operationId": "RoleController_update",
        "summary": "角色管理-修改",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "角色管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/system/role/list": {
      "get": {
        "operationId": "RoleController_findAll",
        "summary": "角色管理-列表",
        "parameters": [
          {
            "name": "pageNum",
            "required": true,
            "in": "query",
            "description": "当前分页",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "orderByColumn",
            "required": false,
            "in": "query",
            "description": "排序字段",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isAsc",
            "required": false,
            "in": "query",
            "description": "排序规则",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "角色管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/system/role/deptTree/{id}": {
      "get": {
        "operationId": "RoleController_deptTree",
        "summary": "角色管理-部门树",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "角色管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/system/role/{id}": {
      "get": {
        "operationId": "RoleController_findOne",
        "summary": "角色管理-详情",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "角色管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      },
      "delete": {
        "operationId": "RoleController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "角色管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/system/role/dataScope": {
      "put": {
        "operationId": "RoleController_dataScope",
        "summary": "角色管理-数据权限修改",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "角色管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/system/role/changeStatus": {
      "put": {
        "operationId": "RoleController_changeStatus",
        "summary": "角色管理-停用角色",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "角色管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/system/role/authUser/allocatedList": {
      "get": {
        "operationId": "RoleController_authUserAllocatedList",
        "summary": "角色管理-角色已分配用户列表",
        "parameters": [
          {
            "name": "pageNum",
            "required": true,
            "in": "query",
            "description": "当前分页",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "orderByColumn",
            "required": false,
            "in": "query",
            "description": "排序字段",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isAsc",
            "required": false,
            "in": "query",
            "description": "排序规则",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userName",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "phonenumber",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AllocatedListDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "角色管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/system/role/authUser/unallocatedList": {
      "get": {
        "operationId": "RoleController_authUserUnAllocatedList",
        "summary": "角色管理-角色未分配用户列表",
        "parameters": [
          {
            "name": "pageNum",
            "required": true,
            "in": "query",
            "description": "当前分页",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "orderByColumn",
            "required": false,
            "in": "query",
            "description": "排序字段",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isAsc",
            "required": false,
            "in": "query",
            "description": "排序规则",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userName",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "phonenumber",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AllocatedListDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "角色管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/system/role/authUser/cancel": {
      "put": {
        "operationId": "RoleController_authUserCancel",
        "summary": "角色管理-解绑角色",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthUserCancelDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "角色管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/system/role/authUser/cancelAll": {
      "put": {
        "operationId": "RoleController_authUserCancelAll",
        "summary": "角色管理-批量解绑角色",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthUserCancelAllDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "角色管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/system/role/authUser/selectAll": {
      "put": {
        "operationId": "RoleController_authUserSelectAll",
        "summary": "角色管理-批量绑定角色",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthUserSelectAllDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "角色管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/system/role/export": {
      "post": {
        "operationId": "RoleController_export",
        "summary": "导出角色管理xlsx文件",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "角色管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Nest-Admin",
    "description": "Nest-Admin 接口文档",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://nest-admin.dooring.vip/prod-api"
    }
  ],
  "components": {
    "securitySchemes": {
      "Authorization": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header",
        "name": "Authorization",
        "description": "请在请求头中携带 JWT 令牌，格式：Bearer <token>"
      }
    },
    "schemas": {
...
      "DateParamsDTO": {
        "type": "object",
        "properties": {}
      },
...
      "CreateRoleDto": {
        "type": "object",
        "properties": {
          "roleName": {
            "type": "string"
          },
          "roleKey": {
            "type": "string"
          },
          "roleSort": {
            "type": "number"
          },
          "status": {
            "type": "string"
          },
          "dataScope": {
            "type": "string"
          },
          "remark": {
            "type": "string"
          }
        },
        "required": [
          "roleName",
          "roleKey",
          "roleSort"
        ]
      },
      "ListRoleDto": {
        "type": "object",
        "properties": {
          "pageNum": {
            "type": "number",
            "description": "当前分页",
            "default": 1
          },
          "pageSize": {
            "type": "number",
            "description": "每页数量",
            "default": 10
          },
          "params": {
            "description": "时间范围",
            "allOf": [
              {
                "$ref": "#/components/schemas/DateParamsDTO"
              }
            ]
          },
          "orderByColumn": {
            "type": "string",
            "description": "排序字段"
          },
          "isAsc": {
            "type": "string",
            "description": "排序规则"
          }
        },
        "required": [
          "pageNum",
          "pageSize"
        ]
      },
      "UpdateRoleDto": {
        "type": "object",
        "properties": {
          "roleName": {
            "type": "string"
          },
          "roleKey": {
            "type": "string"
          },
          "roleSort": {
            "type": "number"
          },
          "status": {
            "type": "string"
          },
          "dataScope": {
            "type": "string"
          },
          "remark": {
            "type": "string"
          },
          "roleId": {
            "type": "number"
          }
        },
        "required": [
          "roleName",
          "roleKey",
          "roleSort",
          "roleId"
        ]
      },
      "ChangeStatusDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "userId",
          "status"
        ]
      },
      "AllocatedListDto": {
        "type": "object",
        "properties": {
          "pageNum": {
            "type": "number",
            "description": "当前分页",
            "default": 1
          },
          "pageSize": {
            "type": "number",
            "description": "每页数量",
            "default": 10
          },
          "params": {
            "description": "时间范围",
            "allOf": [
              {
                "$ref": "#/components/schemas/DateParamsDTO"
              }
            ]
          },
          "orderByColumn": {
            "type": "string",
            "description": "排序字段"
          },
          "isAsc": {
            "type": "string",
            "description": "排序规则"
          },
          "userName": {
            "type": "string"
          },
          "phonenumber": {
            "type": "string"
          },
          "roleId": {
            "type": "string"
          }
        },
        "required": [
          "pageNum",
          "pageSize"
        ]
      },
      "AuthUserCancelDto": {
        "type": "object",
        "properties": {
          "roleId": {
            "type": "number"
          },
          "userId": {
            "type": "number"
          }
        },
        "required": [
          "roleId",
          "userId"
        ]
      },
      "AuthUserCancelAllDto": {
        "type": "object",
        "properties": {
          "roleId": {
            "type": "number"
          },
          "userIds": {
            "type": "string"
          }
        },
        "required": [
          "roleId",
          "userIds"
        ]
      },
      "AuthUserSelectAllDto": {
        "type": "object",
        "properties": {
          "roleId": {
            "type": "number"
          },
          "userIds": {
            "type": "string"
          }
        },
        "required": [
          "roleId",
          "userIds"
        ]
      },
...
    }
  }
}